Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2065 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUVision++/TU/v/Vision++.h
@@ -69,7 +69,7 @@ class BrepCanvasPane : public CanvasPane
     enum DrawMode	{Erase, Draw, Highlight1, Highlight2, Highlight3};
 
     BrepCanvasPane(Window& parentWindow,
-		   u_int width, u_int height, Brep::Root* root)		;
+		   size_t width, size_t height, Brep::Root* root)	;
     ~BrepCanvasPane()							;
     
     virtual void	callback(CmdId id, CmdVal val)			;
